vzcpucheck: print error if VE0 CPU units not set (bug #249)
The problem is if VE0 CPU units is not set (e.g. /etc/init.d/vz not started)
the utility prints bogus information that the node is overcommitted.

To fix, we check that VE0 cpuweight is not 1, and print an appropriate error.

Potential problem is VE0 cpuweight can be explicitly set to 1
(for example, by setting VE0CPUUNITS=500000 in /etc/vz/vz.conf) -- but this
is wrong setting anyway. Still, we print "not set or set too high"
to distinguish between those two cases.

http://bugzilla.openvz.org/249