etc/Makefile.am: do not need to use builddir
As I get it make is always in builddir, so no need to specify it explicitly.
More to say, when you run ./configure in $top_srcdir, $builddir is not set
and thus construction like '$(INSTALL_DATA) $(builddir)/$$file' tries to get
the file from /, not from the current directory, which does not work.