vzmigrate: add --delete to rsync calls
During testing and debugging of a problem on Debian systems
(which I later found out was corrected in 3.0.13 version) I found
that the rsync commands in vzmigrate can give quite unexpected
results.

The command looks basically like this:
rsync -aH ${from} ${to}

This means that files are copied from ${from} to ${to}, but if
files already exist on ${to} from some older copy the old files
that do not exist on ${from} will not be deleted.

This is a problem if --keep-dst, -r no or --online option is used and
in some other situations as well.

There are two places where the rsync command is used for transfer
of filesystem data.

The first one is right after 'log 1 "Syncing private"'. If --keep-dst
is used this is a problem. The solution is to add the --delete option
to rsync.

The second one is right after 'log 2 "Syncing private (2nd pass)"'
and it is a problem if --online option is used and a file has been
deleted between this run and the first rsync run.
The solution is to add the --delete option to rsync.

http://bugzilla.openvz.org/427

[kir@: redo for vzctl HEAD, split lines to obey 80-cols width]