vzctl_veth.h: fix license blurb
This file was taken from OpenVZ kernel sources, so it contained a reference
to copying.SWsoft file (which contains GNU GPL. Apart from the OpenVZ kernel
this reference does not make sense, so replace it with the standard GNU GPL
blurb.

Also, change SWsoft->Parallels and update copyright year while we're at it.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>