Support for udev static devices in container
When a device access is granted to a container, we create an appropriate
device file in container's /dev. The problem is /dev can be under tmpfs
and managed by udev, so device file disappears after container reboot.

The solution is to put the device file in /lib/udev/devices (or
/etc/udev/devices if previous doesn't exist). From there it will be copied
to new /dev on system boot.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>