More strict permissions for /vz/private and /vz/root
As pointed out in bug #954 (and has been generally known for ages)
ordinary users in CT0 can access CT files having the same GIDs/UIDs.
This can be effectively prevented by using mode 700 for /vz/root
and /vz/private. It was already done so in vzctl.spec (for RPMs)
but not in "make install" target. Fix it here as well.

http://bugzilla.openvz.org/954

Signed-off-by: Kir Kolyshkin <kir@openvz.org>