vzctl create: fix loosing CT config file
In case --config option is specified to vzctl create, CT config file
is overwritten with the specified one. First, this is plainly wrong --
if CT config exists, it should be preserved. Second, there is (at least)
one scenario when CT config can be lost (see OpenVZ bug #775) because of
such behaviour.

So, just bail out with error if using --config option and CT config already
exists.

http://bugzilla.openvz.org/775

Signed-off-by: Kir Kolyshkin <kir@openvz.org>