debian-add_ip.sh: introduce interfaces.{head,tail}
``/etc/vz/dists/scripts/debian-add_ip.sh`` generates the container's
``/etc/network/interfaces`` by appending virtual interface
configuration to the CT's ``/etc/network/interfaces.template``.
Besides the name being quite misleading (it's not a template at all),
there are serious problems if one just wants to enable an additional
interface *after* the other ones (e.g. for setting a new alias on
``venet0`` for private communication with the host).

The attached patch enables the use of ``/etc/network/interfaces.head``
as well as ``/etc/network/interfaces.tail``.  The first one is appended
to the resulting ``interfaces`` file *before* the virtual interface info
(and before the so-called "template"), and the second one is appended
at the end.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=514001

[kir@openvz.org: modify "this is auto-generated" comment accordingly]