Incorrect dependencies in init.d script header
With dependency based boot sequencing, I discovered what I believe is
a bug in the init.d script.  The script have a hard dependency on
sshd, yet the package do not depend on openssh-server.  This make me
believe the sshd dependency should be optional.  The script is also
lacking stop dependencies.  The $local_fs dependency is redundant when
depending on $remote_fs.  The script used to start after $syslog, and
should probably continue to do so.

The hard dependency on sshd make the package fail to install.  See
<URL: http://piuparts.debian.org/sid/fail/vzctl_3.0.23-7.log > for the
test log.  Setting severity to important because this problem will not
affect everyone, only those without openssh-server already installed.

<URL: http://refspecs.freestandards.org/LSB_2.1.0/LSB-generic/LSB-generic/initscrcomconv.html >
documents the LSB header format.  Some debian notes are available from
<URL: http://wiki.debian.org/LSBInitScripts >.

This patch solve the problem.

Signed-off-by: Petter Reinholdtsen <pere@hungry.com>
Acked-by: Kir Kolyshkin <kir@openvz.org>