vzctl enter: fix argv[0] for sh case
vzctl enter doesn't work for a container that use busybox as a shell.
This is because busybox relies on argv[0] as an 'applet' name and
obviously there is no 'bash' applet in it.

Fix is to change argv[0] accordingly before trying to execute /bin/sh.

Bug reported by Floris Bos <openvz-bug@maxnet.eu>

http://bugzilla.openvz.org/1314

Signed-off-by: Kir Kolyshkin <kir@openvz.org>