Fix removing IPs on CT start
This, together with a few previous commits to dist/scripts,
fixes the following bug: when you stop the container and
remove its IP addresses, they are still there after CT start.

This was happening because in case of empty list of IPs
configuration (i.e. running *_add_ip.sh) was skipped.
Do not skip it when we start a container.

Check for op == ADD is added because vps_net_ctl is called
two times -- once with op=DEL and once with op=ADD. Since
the list of IPs is emtpy we don't have anything to delete.

http://bugzilla.openvz.org/478

Signed-off-by: Kir Kolyshkin <kir@openvz.org>