Add new global variable ERROR_ON_ARPFAIL
vzctl have a functionality to check all IPs to be added to a container
for collisions, i.e. for each address it checks if there is such an
address on the network (by sending a broadcast ARP request). If it gets
a positive reply (i.e. there is such an IP already) it prints a warning.

Now with this patch and "ERROR_ON_ARPFAIL" set to "yes" vzctl errors
out in case an IP address is found. This can be unwanted in some
network setups so we keep the old behaviour by default.

Based on patch from Konstantin Pavlov <thresh@altlinux.org>

http://bugzilla.openvz.org/919

Signed-off-by: Kir Kolyshkin <kir@openvz.org>