vzctl set: fix option names
A couple of set ptions had extra characters at the end.

The bug was never ever noticed (since the very first commit
c685a071dc1c2ba4df185bc1f7b27fd090837242 dated Aug 14 2005!)
because GNU getopt allows to shorten option names (if there
is no ambiguity) so --oomguarpages and --oomguarpagess is
considered to be the same option.

Therefore the only manifestation of this bug was an ability
to specify --oomguarpagess and --numothersockk options.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>