src/lib/config.c: use conf_store_strlist()
There are two places which basically reimplement conf_store_strlist().
Remove the duplicated code, use the function.

Since parameters such as IP_ADDRESS have a special case when the
value is empty ("remove all IPs") we need to add a parameter to
conf_store_strlist() which specifies whether empty lists should be
saved or not.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>