vzlist: add 'bootorder' field
This is a second patch in a set to fix OpenVZ bug #1300
(first is f882a75f4887c1d3ab0c2a29a2605187f01189bd).

This patch adds 'bootorder' to the list of fields which vzlist
can show, and documents it in man/vzlist.8.

Note that sorting for bootorder is implemented according to specs,
i.e. then one runs vzlist -s -bootorder, the list is sorted for
bootorder (highest first, lowest and unset last) and then by CTID
(lowest first, highest last). This is somewhat controversial
(read "hackish") but since vzlist can't be told to have two
sorting criteria this is the only simple way to go.

http://bugzilla.openvz.org/1300

Signed-off-by: Kir Kolyshkin <kir@openvz.org>