man/vzlist.8: fix disappeared ONBOOT
Inside a table we can not use two tabs, because the tab is a field
separator. There was two tabs before ONBOOT so it was considered to
be the contents of the third, not the second row. Because of that
it was invisible and groff was complaining about extra contents.

The fix is just to remove one tab.

tbl is so sensitive.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>