vzlist: allow to list stopped CT's cpu params when OpenVZ is not loaded
We want to be able to see values from config files even if OpenVZ is not
running. There is already some code allowing to do that (although you
need to specify -S to use it), but not yet for the case we want to see
CPU-related params. Fix it.

Before this patch:
	$ vzlist -S -octid,cpuunits,cpulimit 101
	Unable to open /proc/vz/veinfo
	Unable to open /proc/fairsched
	$ echo $?
	1

After this patch:
	$ vzlist -S -octid,cpulimit,cpuunits 101
	Unable to open /proc/vz/veinfo
	      CTID  CPULIM  CPUUNI
	       101     101   10100

Signed-off-by: Kir Kolyshkin <kir@openvz.org>