debian-add_ip.sh: do not touch /etc/network/interfaces for no reason
On Debian CT start, we need to prepare proper /etc/network/interfaces
suitable for adding venet devices if we are adding new IPs to it.
But even if we don't, we need to remove old IP addresses that
might left from the previous CT run. This is what the code did.

The problem is if a user has its own network setup (say with veth)
and does not use venet at all we should not touch his
/etc/network/interfaces. This is what this patch is trying to accomplish.

Yet another attempt to fix bug #478:
http://bugzilla.openvz.org/478

Signed-off-by: Kir Kolyshkin <kir@openvz.org>