vzmigrate: check exit code from rsync
the exit code of rsync isn't checked now, because it's in the pipeline,
which is necessary for output parsing that show overall progress. Now this
functionality is broken, because a format of rsync output has been
changed.
http://bugzilla.openvz.org/show_bug.cgi?id=1066

Let's delete this "progress" and will check rsync exit code. it's important,
because useful data may be lost and we have some bugs about it:

http://bugzilla.openvz.org/show_bug.cgi?id=1104
http://bugzilla.openvz.org/show_bug.cgi?id=1006
http://bugzilla.openvz.org/show_bug.cgi?id=879