man/Makefile.am: add target for checking/fixing man page dates
The idea is simple: get the last git commit date for the man page
file and put it into that file, then use git diff to see if there
are any changes.

Note though that this script can not be used to fix those dates,
otherwise last git commit date will become today. So this is only
for checking, not for fixing.

Interestingly enough, only one date was fixed when I run this test,
and this only happened because one commit was committed later than
created.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>