vzgetnetdev(): stricter regexes when looking for device numbers
We are parsing "ip address list" output for lines with interfaces,
they begin with one or more digits and this is catched by the current regex.
But then after the digit(s) comes a colon and a space, so let's include
these two characters to the regex as well to make it stricter.

Again as previous patch this is not a bugfix but just an enhancement.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>