Enable sysfs feature by default
Most of the newer distros need it, so enable it. User can still
disable it explicitly if needed, by using --features sysfs:off.

The feature was not on by default because older kernels used to
pin about 2M of kernel memory per sysfs instance for container.
Now it's not the case so we can enable it harmlessly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>