vzctl start: mount /proc and /sys inside CT
Some of the modern distros rely on initrd scripts doing this,
and since CTs don't have initrd we do it in vzctl.

Another reason is commit 98642b9e7a640a9c989898b69870effcbbfb651e
which tries to mount to /proc/fs/nfsd so we need to have /proc
mounted before.

Also, move mk_quota_link to after mounting /proc and such,
because it involves /proc (as a link destination so it doesn't
really matter but still better this way).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>