vzmigrate: saner output in case of absent config
There is a statement that checks the readability of both
global and CT configs in one sitting. In case of problem
it reports:

   Can't read global config or CT #$VEID config file

This message is a bit strange, because it's not clear what file is
missing.

To fix, remove the check for per-CT config file and fix the error
message accordingly. Check for per-CT config is done later, when
vzctl status is called -- for the absent config case it returns
'deleted' which is acted upon appropriately.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>