vzmigrate: saner output for non-existing CTID
Commit 484afba08 added support for CT names. It works, but
one corner case is not working particularly well. When non-existing
numeric CT ID is supplied, vzlist prints "Container(s) not found"
error which is a bit misleading. (If non-numeric non-existing ID
is supplied to vzlist the error is "CT ID 12434a is invalid"
which is fine in the context of running vzmigrate.)

To fix this problem, dd a check if supplied CTID is a number,
and in case it is, do not try to call vzlist for resolving.

  # vzmigrate otherhost 123
Before:
  Container(s) not found
Now:
  CT #123 doesn't exist

Signed-off-by: Kir Kolyshkin <kir@openvz.org>