vzmigrate: ignore error 24 from the first rsync run
Now, when we really started to check the exit code from rsync, we got
the following problem. Apparently, sometimes first rsync run fails
with error 24, it happens when a file to be transferred is removed.
Since the container is running at this time there is nothing wrong
with it, so ignore error 24.

Should fix bug #1580.

http://bugzilla.openvz.org/1580

Reported-by: Roman Veretelnikov <roman@rover.pp.ru>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>