bash_completion: remove unneeded code
We tried to distinguish between CTID and CT NAME, and used
"vzlist -N ${name}" in the latter case. There is no need to use -N,
vzlist supports CT NAMES natively, so remove this kludge.

BTW, the code that was used to distinguish had an assumption that
CT NAME should contain a [a-zA-Z] character. This is incorrect.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>