Do not allow INT_MAX CT ID
vzctl (and mee, too) thought that INT_MAX makes perfectly valid
CT ID. Unfortunately kernel doesn't agree with us, failing to start
such a container with the following diagnostics:

	Can't create fairsched node 2147483647
	CT: 2147483647: stopped
	CT: 2147483647: failed to start with err=-16

This is because in fair scheduler node with ID of MAX_INT
is reserved for CT0. So check for that CTID and complain that
it's invalid when trying to create it, rather than let it be
created and then failed to start.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>