Fix netmask for IPv6 address in a Gentoo container
The gentoo-add_ip.sh script (vzctl version 3.0.24) doesn't detect if it is
adding an IPv4 or IPv6 address. In /etc/conf.d/net this results in:

config_venet0=" 192.168.13.203/32  2a01:238:43d4:5c00::203:1/32 "

That is wrong for IPv6 and results in a non working IPv6 connection. Gentoo
uses the net-tools package and for some reason a /0 instead of /128 must be
used. I've patched the gentoo-add_ip.sh script to something that works for me.

http://bugzilla.openvz.org/1571