vzctl start: do not remove 127.0.0.1 aliases in /etc/hosts
Commit 164d72740 from OpenVZ bug #855 fixes a lot of issues with
vzctl way of handling containers' /etc/hosts. Unfortunately the patch
was incomplete and one old issue was not fixed. The issue is when a CT
root modifies the /etc/hosts line about 127.0.0.1 a.k.a. localhost,
these modifications are lost when vzctl updates it. The issue was reported
again as Debian bug #594003, then OpenVZ bug #1618.

This commit adds the missing hunk, hopefully fixing the issue.

http://bugzilla.openvz.org/1618
http://bugs.debian.org/594003
http://bugzilla.openvz.org/855

Reported-by: Stefan Alfredsson <alfs@debian.org>
Signed-off-by: Marcin Owsiany <marcin@owsiany.pl>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>