init.d/vzevent: support Debian-specific file to set vzeventd options
For Red Hat like distros and SUSE, options for various daemons and such
are usually set in files residing in /etc/sysconfig, in our case
this is /etc/sysconfig/vzeventd.

Debian uses /etc/default for such stuff, so check if /etc/default/vzeventd
exists and is readable and source it.

For example, if you want vzeventd to be more verbose, do
  echo 'OPTIONS=-v' > /etc/default/vzeventd	# Debian
  echo 'OPTIONS=-v' > /etc/sysconfig/zeventd	# Red Hat, Fedora, SUSE
and then /etc/init.d/vzeventd restart

Signed-off-by: Kir Kolyshkin <kir@openvz.org>