vzctl set --diskquota: do not overwrite command line with global parameter
Previous commit adds --diskquota option. Now we have this option in
* cmd_p (parameters obtained from command line)
* gparam (parameters from global config)
* vps_p (parameters from CT config)

For some reason (unknown to me) function merge_global_config
put the dq.enable parameter from global ones and set it into
cmd_p -- thus overwriting what we got from the command line.

The fix is to remove that code. I was not able to find out why
it was needed there, but some code reading and testing showed
that removing it makes no harm. I sincerely hope so.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>