vzctl quotaoff: fix wrong exit code if quota is not running
If quota is not running, we should not return an error
when trying to turn it off. There is such a code already,
but it forgot to set ret to zero, therefore the native
exit code from vzquota (6) was returned by vzctl which
makes no sense (for vzctl 6 means "Not enough system
resources").

Fix it by 'ret=0' in case quota is not running.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>