quota_ctl(), vps_set_quota(): refactor/fix
There is no need to do hairy trickery with return values
in quota_ctl() -- always return the exit code of vzquota.

In vps_set_quota(), when we run quota_ctl(STAT2), we need
to find out if UGID quota is running or not. If it's not
running vzquota returns exit code of 9, which we check for.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>