etc/init.d/vz*: add kludge to modify default CT config
On a vswap-enabled kernel, it makes great sense to use vswap-enabled
CT configuration by default. For the sake of better solution, let's
check in vz initscript whether we run vswap-enabled kernel and if yes,
modify the value of CONFIGFILE= parameter in /etc/vz/vz.conf to be
'vswap-256m' instead of 'basic'. If CONFIGFILE is modified, do nothing.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>