vzmemcheck: only account for running containers.
While it is usually not the case, sometimes /proc/user_beancounters contains
values from the old (not currently running) containers. There are a few reasons
why it can be the case, with the most common being a kernel bug.

Nevertheless, vzmemcheck should not use beancounters for not running CTs.
We do so by getting a list of running CTs first, and then for every UBC
check that such UBC ID exist in a list of running CTs.

Should fix bug #1757.

http://bugzilla.openvz.org/1757

Reported-by: Pete de Zwart <pdzwart@atlassian.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>

Signed-off-by: Kir Kolyshkin <kir@openvz.org>